pyproject.toml
[tool.poetry]
name = "django-adminlte4"
version = "0.1.0"
description = ""
authors = ["Adam Charnock <adam@adamcharnock.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
